requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
firebase-admin==5.0.0
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
